+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Resource data will be available with the resource object. This enhancement helps
 - SAS interconnect
 - SAS interconnect type
 - SAS logical interconnect
+- Server hardware
 - Server hardware type
 - Server profile template
 - Switch type